优化供应商运费接口

+++ dangjia-subgroup/dangjia-config/dangjia-common/src/main/java/com/dangjia/acg/modle/supplier/DjSupplier.java
@@ -10,6 +10,7 @@ import lombok.experimental.FieldNameConstants;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.Table;
+import java.math.BigDecimal;
 
 /**
  * Created with IntelliJ IDEA.
@@ -92,6 +93,6 @@ public class DjSupplier extends BaseEntity {
     @Column(name = "transportation_cost")
     @Desc(value = "每单收取运费")
     @ApiModelProperty("每单收取运费")
-    private Double transportationCost;
+    private BigDecimal transportationCost;
 
 }
